+++ core/llm/prompt_template_manager.py
@@ -1,0 +1,61 @@
+class PromptTemplateManager:
+    """프롬프트 템플릿을 중앙에서 관리하는 클래스"""
+    
+    @staticmethod
+    def get_video_summary_prompt() -> str:
+        """유튜브 영상 요약용 프롬프트 템플릿"""
+        return """
+당신은 유튜브 영상 자막을 분석해 구간별 개요를 작성하는 AI입니다.
+자막은 실제 영상의 전체 내용을 포함하고 있으며, 당신의 목표는 이 자막을 기반으로 10초 단위로 영상의 흐름을 정리하는 것입니다.
+- 각 소제목은 영상의 주제를 대표해야 합니다.
+- 각 구간은 영상 흐름을 반영하여 약 10초 단위(±5초 허용)로 구분해야 합니다.
+- 출력은 반드시 위의 예시 형식을 따르며, 불필요한 부가 설명 없이 개요만 출력해야 합니다.
+- 각 구간의 제목과 시간 범위를 표시하세요.
+- 각 구간 설명은 3인칭 시점으로, 객관적으로 간결하게 서술하세요.
+- 1인칭 표현(예: "나는", "내가") 대신 "화자", "그/그녀", "인터뷰어" 등의 표현을 사용하세요.
+- 각 구간별 핵심 내용은 최대 3문장 이내로 요약하세요.
+- 전체 개요를 I, II, III 등으로 번호 매겨 구성하세요.
+- 불필요한 로그나 시스템 메시지를 출력하지 마세요
+
+다음과 같은 형식으로 결과를 반드시 출력해야 합니다:
+
+다음은 제공된 영상 스크립트를 기반으로 한 유튜브 영상 개요의 번역입니다:
+요구 사항:
+
+
+I. 도입 (0:00 - 0:25)
+화자가 자신을 대학생이라고 소개한다.  
+본인의 채널 및 콘텐츠에 대해 간단히 설명한다.
+
+II. 글쓰기와 재능 (0:25 - 0:59)  
+화자는 글쓰기를 좋아한다고 말한다.  
+인터뷰어는 화자에게 리더십 같은 다른 재능도 있다고 언급한다.  
+화자는 자신의 리더십 능력을 과소평가한다.
+
+...
+
+질문: {input}
+문서 내용: {context}
+답변:""".strip()
+    
+    @staticmethod
+    def get_comment_reaction_prompt() -> str:
+        """댓글 반응 분석용 프롬프트 템플릿"""
+        return """
+당신은 유튜브 댓글을 분석하는 AI입니다.
+댓글의 감정, 주요 키워드, 시청자 반응을 분석해주세요.
+
+질문: {input}
+문서 내용: {context}
+답변:""".strip()
+    
+    @staticmethod
+    def get_video_evaluation_prompt() -> str:
+        """비디오 평가용 프롬프트 템플릿"""
+        return """
+당신은 유튜브 비디오를 분석하는 AI입니다.
+비디오의 특성, 콘텐츠 트렌드, 성장 요인을 분석해주세요.
+
+질문: {input}
+문서 내용: {context}
+답변:""".strip()
